backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
boto3==1.25.5
botocore==1.28.5
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-oauth-toolkit==2.1.0
django-storages==1.13.1
djangorestframework==3.13.1
drf-social-oauth2==1.2.1
drf-spectacular==0.24.2
ecdsa==0.18.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.16.0
jwcrypto==1.3.1
numpy==1.23.2
oauthlib==3.2.0
pandas==1.4.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
